% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{contingency_table}
\alias{contingency_table}
\title{Contingency table}
\usage{
contingency_table(...)
}
\arguments{
\item{...}{Passed to table}
}
\value{
Table object
}
\description{
Include NAs and add totals to table.
}
