% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_matrix.R
\name{cosine_simi}
\alias{cosine_simi}
\title{Cosine similarity between vectors and/or matrices.}
\usage{
cosine_simi(x, y)
}
\arguments{
\item{x}{Numeric vector or matrix}

\item{y}{Numeric vector or matrix. If missing, copied from parameter x.}
}
\value{
Symmetric numeric similarity matrix
}
\description{
Inputs will be L2 normalized, then matrix multiplied.
If second input is missing, first input will be recycled, which enables to
efficiently compute cosine similarities between the rows of a rectangular
matrix.
}
