% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_matrix.R
\name{dist_matrix}
\alias{dist_matrix}
\title{dist_matrix}
\usage{
dist_matrix(data, method = "euclidean", n_cores = 1)
}
\arguments{
\item{data}{Rectangular numeric matrix [Observations, Features]}

\item{method}{Methods accepted by amap::Dist or cosine and norm_inprod}

\item{n_cores}{Number of cores}
}
\value{
Distance symmetric matrix
}
\description{
Dispatch of amap::Dist, cosine_dist, and norm_inprod methods.
}
