% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_models.R
\name{ensemble_metrics}
\alias{ensemble_metrics}
\title{Compute ensemble metrics}
\usage{
ensemble_metrics(
  n_top = 0,
  df_params,
  metrics = NULL,
  metrics_exclude = NULL,
  n_models = 10
)
}
\arguments{
\item{n_top}{Threshold of number of models to rank}

\item{df_params}{Output of opticskxi_pipeline}

\item{metrics}{Names of metrics to use. Any of those computed by
opticskxi_pipeline, e.g. 'sindex', 'ch', 'dunn',
'dunn2', 'widestgap', 'entropy' etc.
NULL for all (8).}

\item{metrics_exclude}{Names of metrics to exclude.
Typically used with metrics = NULL. E.g. 'entropy'.}

\item{n_models}{Number of best models to return}
}
\value{
List of metrics matrix and df_params subsetted to best models
}
\description{
Use models' rankings over several metrics to select best model.
}
