% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_models.R
\name{ensemble_models}
\alias{ensemble_models}
\title{Select models based on ensemble metrics}
\usage{
ensemble_models(l_ensemble_metrics, n_models = 4)
}
\arguments{
\item{l_ensemble_metrics}{Output of function ensemble_metrics}

\item{n_models}{Number of best models to return}
}
\value{
List of parameters of best models
}
\description{
Typically we will call ensemble_metrics with varying numbers of ranks to
consider and this function will sum up the ranks from those calls.
}
