% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimred.R
\name{fortify_dimred}
\alias{fortify_dimred}
\title{Fortify a dimension reduction object}
\usage{
fortify_dimred(
  m_dimred,
  m_vars = NULL,
  v_variance = NULL,
  sup_vars = NULL,
  var_digits = 1
)
}
\arguments{
\item{m_dimred}{Projection matrix}

\item{m_vars}{Rotation matrix (optional)}

\item{v_variance}{Explained variance (optional)}

\item{sup_vars}{Optional supplementary variables}

\item{var_digits}{Explained variance percent digits}
}
\value{
Data frame
}
\description{
Fortify a dimension reduction object
}
\examples{
pca <- prcomp(iris[-5])
df_pca <- fortify_dimred(pca$x)
}
\seealso{
\link{fortify_pca}, \link{fortify_ica}
}
