% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimred.R
\name{fortify_ica}
\alias{fortify_ica}
\title{Get and fortify ICA}
\usage{
fortify_ica(m_data, ..., sup_vars = NULL)
}
\arguments{
\item{m_data}{Input matrix}

\item{...}{Passed to fastICA::fastICA}

\item{sup_vars}{Optional supplementary variables}
}
\value{
Fortified dimension reduction
}
\description{
Get and fortify ICA
}
\examples{
df_ica <- fortify_ica(iris[-5], n.comp = 2)
}
\seealso{
\link{fortify_dimred}, \link{fortify_pca}
}
