% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimred.R
\name{fortify_pca}
\alias{fortify_pca}
\title{Get and fortify PCA}
\usage{
fortify_pca(m_data, ..., sup_vars = NULL)
}
\arguments{
\item{m_data}{Input matrix}

\item{...}{Passed to stats::prcomp}

\item{sup_vars}{Optional supplementary variables}
}
\value{
Fortified dimension reduction
}
\description{
Get and fortify PCA
}
\examples{
df_pca <- fortify_pca(iris[-5])
df_pca <- fortify_pca(iris[-5], sup_vars = iris[5])
}
\seealso{
\link{fortify_dimred}, \link{fortify_ica}
}
