% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opticskxi_pipeline.R
\name{get_best_kxi}
\alias{get_best_kxi}
\title{Get best k-Xi model}
\usage{
get_best_kxi(df_kxi, metric = "avg.silwidth", rank = 1)
}
\arguments{
\item{df_kxi}{Data frame returned by opticsxi_pipeline}

\item{metric}{Metric to choose best model}

\item{rank}{Rank(s) of model to choose, ordered by decreasing metric}
}
\value{
df_kxi row with specified metric and rank,
        simplified to a list if only one rank selected
}
\description{
Select k-Xi clustering model based on a metric and a rank
}
\seealso{
\link{opticskxi_pipeline}
}
