% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optics_plots.R
\name{gtable_kxi_profiles}
\alias{gtable_kxi_profiles}
\title{Gtable OPTICS k-Xi distance profiles}
\usage{
gtable_kxi_profiles(df_kxi, metric = "avg.silwidth", rank = 1:4, ...)
}
\arguments{
\item{df_kxi}{Data frame returned by opticskxi_pipeline}

\item{metric}{Metric to choose best clustering model}

\item{rank}{Ranks of models to plot, ordered by decreasing model metric}

\item{...}{Passed to ggplot_kxi_profile}
}
\description{
Plot OPTICS distance profiles of k-Xi clustering models
}
\seealso{
\link{opticskxi_pipeline}
}
