% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_psych_embeds.R
\docType{data}
\name{m_psych_embeds}
\alias{m_psych_embeds}
\title{A dataset containing the embeddings matrix of psychological related words}
\format{
A matrix with 831 words in rows and 100 embedding dimensions in
columns.
}
\usage{
data("m_psych_embeds")
}
\description{
Data containing Glove embeddings of psychological related
words, useful for demonstrating the use of ensemble metrics.
}
\details{
The dataset contains groups of related words among other irrelevant words.
}
