% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_psychwords.R
\docType{data}
\name{m_psychwords}
\alias{m_psychwords}
\title{A dataset containing words by embeddings matrix}
\format{
A matrix with 799 words in rows and 100 embedding dimensions in
columns.
}
\usage{
data("m_psychwords")
}
\description{
Data containing Glove embeddings of psychological related
words, useful for demonstrating the use of the modified opticskxi pipeline
psychkxi.
}
\details{
The dataset contains 2 main hierarchical clusters and each has subclusters.
}
