% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nice_palette}
\alias{nice_palette}
\title{Nice palette}
\usage{
nice_palette(groups, rainbow = FALSE)
}
\arguments{
\item{groups}{Vector, each unique value will get a color}

\item{rainbow}{If TRUE, rainbow-like colors, else differentiate successive
values}
}
\value{
Vector of colors
}
\description{
Color palette
}
