% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_matrix.R
\name{norm_inprod}
\alias{norm_inprod}
\title{norm_inprod}
\usage{
norm_inprod(m)
}
\arguments{
\item{m}{Numeric matrix}
}
\value{
Numeric matrix
}
\description{
Normalized inner product with transposed input matrix
}
