% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_matrix.R
\name{normalize}
\alias{normalize}
\title{Matrix normalization}
\usage{
normalize(m, norm = c("l1", "l2", "none"))
}
\arguments{
\item{m}{\code{matrix} (sparse or dense).}

\item{norm}{\code{character} the method used to normalize term vectors}
}
\value{
normalized matrix
}
\description{
normalize matrix rows using given norm. Copied from text2vec
}
