% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_table}
\alias{print_table}
\title{Print table}
\usage{
print_table(table_obj, label)
}
\arguments{
\item{table_obj}{Table object}

\item{label}{Latex label}
}
\value{
None
}
\description{
Print knitr::kable latex table with legend at bottom.
}
