% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_vignette_table}
\alias{print_vignette_table}
\title{Print vignette table}
\usage{
print_vignette_table(table_obj, label)
}
\arguments{
\item{table_obj}{Table object}

\item{label}{Latex label}
}
\value{
None, side-effect prints a Latex table
}
\description{
Print knitr::kable latex table with legend at bottom.
}
