% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{residuals_table}
\alias{residuals_table}
\title{Residuals table}
\usage{
residuals_table(...)
}
\arguments{
\item{...}{Passed to contingency_table and chisq.test}
}
\value{
Matrix
}
\description{
Bind contingency table and Pearson Chi-squared residuals.
}
