% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_matrix.R
\name{stddev_mean}
\alias{stddev_mean}
\title{stddev_mean}
\usage{
stddev_mean(m)
}
\arguments{
\item{m}{Numeric matrix}
}
\value{
Mean of standard deviations of matrix columns
}
\description{
Get mean of standard deviations of matrix columns
}
