% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_psychwords.R
\docType{data}
\name{m_psychwords}
\alias{m_psychwords}
\title{A dataset containing words by embeddings matrix}
\format{
A matrix with 800 words in rows and 100 embedding dimensions in
columns.
}
\usage{
data("m_psychwords")
}
\description{
Data containing Glove embeddings of psychological related
words, useful for demonstrating the use of the modified opticskxi pipeline
psychkxi.
}
\details{
The dataset contains 2 main hierarchical clusters (each has subclusters).
}
\examples{
data('m_psychwords')
df_params = expand.grid(n_xi = 9:10, pts = c(15, 20), dist = 'cosine',
  dim_red = 'ICA', n_dimred_comp = c(10, 15))
df_kxi = opticskxi:::psych_kxi_ensemble_models(m_psychwords, df_params)

}
